requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.3.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
cymem==2.0.7
Cython==0.29.33
distlib==0.3.6
Django==4.1.6
django-crispy-forms==1.14.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
exceptiongroup==1.1.0
filelock==3.9.0
flake8==5.0.4
hypothesis==6.68.1
identify==2.5.17
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.2
mccabe==0.7.0
ml-datasets==0.2.0
mock==2.0.0
murmurhash==1.0.9
mypy==1.0.0
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
packaging==23.0
pathspec==0.11.0
pathy==0.10.1
pbr==5.11.1
platformdirs==3.0.0
pluggy==1.0.0
pre-commit==3.0.4
preshed==3.0.8
pycodestyle==2.9.1
pydantic==1.10.4
pyflakes==2.5.0
pytest==7.2.1
pytest-timeout==1.4.2
PyYAML==6.0
requests==2.28.2
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
-e git+https://github.com/explosion/spaCy@2d4fb94ba0a23523cc9adb65e0dcf92bbf6177b6#egg=spacy
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.7
tomli==2.0.1
tqdm==4.64.1
typer==0.7.0
types-docutils==0.19.1.3
types-mock==5.0.0.2
types-requests==2.28.11.12
types-setuptools==67.2.0.1
types-urllib3==1.26.25.5
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.19.0
wasabi==1.1.1
